Removed the nil v NSNull problem (for the currentGameID, where nil was ‘undefined, ever’ and null was ‘has been in game but no longer is’) Also fixed compiler warning on NSInteger in string format.

+++ CycleOfViolence/CycleOfViolence/COVInactiveGameViewController.m
@@ -55,8 +55,8 @@
     } else {
         NSInteger days = (int) ( timeToStart / (3600*24) );
         NSInteger hours = ( (int) timeToStart % (3600*24) ) / 3600;
-        self.countdown.text = [NSString stringWithFormat:@"Starting in %d days, %d hours.",
-                               days, hours];
+        self.countdown.text = [NSString stringWithFormat:@"Starting in %ld days, %ld hours.",
+                               (long)days, (long)hours];
     }
     
     // Show the buttons selectively.
